Add manage users and confirm when delete

+++ batdongsan/src/main/webapp/WEB-INF/views/admin/News/listNews.jsp
@@ -113,9 +113,9 @@
 
 										</a> <a href='listNews/hide/${n.newsId}.html' class='button-item'>
 											<i class="fa-regular fa-eye-slash"></i> <span>Ẩn tin</span>
-										</a> <a href='listNews/delete/${n.newsId}.html'
-											class='button-item'> <i class="fa-regular fa-trash-can"></i> <span>Xóa
-												tin</span>
+										</a> <a href='listNews/delete/${n.newsId}.html' onclick="return confirmDelete();"
+											class='button-item'> <i class="fa-regular fa-trash-can"></i>
+											<span>Xóa tin</span>
 										</a> <a href='listNews/update/${n.newsId}.html'
 											class='button-item'> <i class='fa-solid fa-pencil'></i> <span>Sửa
 												tin</span>
@@ -189,6 +189,9 @@
 								window.location.href = url;
 							}
 						})
+		function confirmDelete() {
+			return confirm("Bạn có chắc chắn muốn xóa danh mục này không?");
+		}
 	</script>
 </body>
 </html>
